- moved restart file time stack to its own file for clarity
