+++ src/roguewave/wavewatch3/__init__.py
@@ -1,4 +1,5 @@
-from roguewave.wavewatch3.restart_file import RestartFile, RestartFileTimeStack
+from roguewave.wavewatch3.restart_file import RestartFile
+from roguewave.wavewatch3.restart_file_time_stack import RestartFileTimeStack
 from roguewave.wavewatch3.io import (
     clone_restart_file,
     open_restart_file,
